requirements.txt
tensorflow
matplotlib
numpy
pillow